test: consolidate acpshim deadlock regression tests
